Extract plotting room repository slice
