+++ changelog.d/+plotting-repository-slice.changed.md
@@ -1,0 +1,2 @@
+Moved plotting room and plotting room participant persistence into a dedicated
+repository slice while preserving the forum repository facade.
