+++ pyproject.toml
@@ -96,6 +96,7 @@ ignore = [
 "src/elbysodic/db/repositories/facets.py" = ["S608"]
 "src/elbysodic/db/repositories/materials.py" = ["S608"]
 "src/elbysodic/db/repositories/plot_hooks.py" = ["S608"]
+"src/elbysodic/db/repositories/plotting.py" = ["S608"]
 "src/elbysodic/db/repositories/wanted.py" = ["S608"]
 
 # =============================================================================
